so many updates prod deletion etc

+++ pages/admin/product/update/[id].js
@@ -1,7 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import { useGetProductByIdQuery, useUpdateProductMutation, useDeleteProductMutation, useUploadProductImageMutation } from '../../../../redux/api/productApiSlice';
+import { useRemoveAllOfProductFromAllofCartMutation } from '../../../../redux/api/cartApiSlice';
 import { useFetchCategoriesQuery } from '../../../../redux/api/categoryApiSlice';
+import { useRemoveFromAllWishListMutation } from '../../../../redux/api/wishlistApiSlice';
+import { toast } from 'react-toastify';
 
 const ProductUpdate = () => {
     const router = useRouter();
@@ -15,7 +18,6 @@ const ProductUpdate = () => {
         skip: !productId,
     });
 
-    const [image, setImage] = useState("");
     const [name, setName] = useState("");
     const [description, setDescription] = useState("");
     const [price, setPrice] = useState("");
@@ -24,33 +26,85 @@ const ProductUpdate = () => {
     const [stock, setStock] = useState("");
     const [offers, setOffers] = useState("");
     const [returnpolicy, setReturnPolicy] = useState("");
+    const [frontImage, setFrontImage] = useState(null);
+    const [backImage, setBackImage] = useState(null);
+    const [frontDesign, setFrontDesign] = useState(null);
+    const [backDesign, setBackDesign] = useState(null);
+    const [images, setImages] = useState([]);
+    const [frontImageUrl, setFrontImageUrl] = useState(null);
+    const [backImageUrl, setBackImageUrl] = useState(null);
+    const [frontDesignUrl, setFrontDesignUrl] = useState(null);
+    const [backDesignUrl, setBackDesignUrl] = useState(null);
+    const [imagesUrl, setImagesUrl] = useState([]);
+    const [frontImageUrlToDisplay, setFrontImageUrlToDisplay] = useState(null);
+    const [backImageUrlToDisplay, setBackImageUrlToDisplay] = useState(null);
+    const [frontDesignUrlToDisplay, setFrontDesignUrlToDisplay] = useState(null);
+    const [backDesignUrlToDisplay, setBackDesignUrlToDisplay] = useState(null);
+    const [imagesUrlToDisplay, setImagesUrlToDisplay] = useState([]);
 
     const { data: categories = [] } = useFetchCategoriesQuery();
 
     const [uploadProductImage] = useUploadProductImageMutation();
     const [updateProduct] = useUpdateProductMutation();
     const [deleteProduct] = useDeleteProductMutation();
+    const [removeAllOfProductFromAllofCart] = useRemoveAllOfProductFromAllofCartMutation();
+    const [removeFromAllWishlist] = useRemoveFromAllWishListMutation();
 
     useEffect(() => {
         if (productData) {
             setName(productData.name);
             setDescription(productData.description);
             setPrice(productData.price);
-            setCategory(productData.category?._id || "");
+            setCategory(productData.category);
             setGender(productData.gender);
-            setImage(productData.image);
             setStock(productData.countInStock);
             setOffers(productData.offers);
             setReturnPolicy(productData.returnpolicy);
+            setFrontImageUrl(productData.frontImage);
+            setBackImageUrl(productData.backImage);
+            setFrontDesignUrl(productData.frontDesign);
+            setBackDesignUrl(productData.backDesign);
+            setImagesUrl(productData.images);
+            setFrontImageUrlToDisplay(productData.frontImage);
+            setBackImageUrlToDisplay(productData.backImage);
+            setFrontDesignUrlToDisplay(productData.frontDesign);
+            setBackDesignUrlToDisplay(productData.backDesign);
+            setImagesUrlToDisplay(productData.images);
+            console.log('productData:', productData);
         }
     }, [productData]);
 
-    const uploadFileHandler = async (e) => {
+    useEffect(() => {
+        if (productData) {
+            console.log('Product category from productData:', productData.category);
+            setCategory(productData.category); // This should set the category correctly
+        }
+    }, [productData]);
+
+    useEffect(() => {
+        console.log('Updated category state:', category);
+    }, [category]);
+
+
+    const uploadImages = async (e) => {
         const formData = new FormData();
-        formData.append("image", e.target.files[0]);
+        if (frontImage) formData.append("frontImage", frontImage);
+        if (backImage) formData.append("backImage", backImage);
+        if (frontDesign) formData.append("frontDesign", frontDesign);
+        if (backDesign) formData.append("backDesign", backDesign);
+        if (images.length) {
+            images.forEach((image) => {
+                formData.append("images", image);
+            });
+        }
         try {
             const res = await uploadProductImage(formData).unwrap();
-            setImage(res.image);
+            if (res.imageUrls?.frontImage?.[0]) setFrontImageUrl(res.imageUrls.frontImage[0]);
+            if (res.imageUrls?.backImage?.[0]) setBackImageUrl(res.imageUrls.backImage[0]);
+            if (res.imageUrls?.frontDesign?.[0]) setFrontDesignUrl(res.imageUrls.frontDesign[0]);
+            if (res.imageUrls?.backDesign?.[0]) setBackDesignUrl(res.imageUrls.backDesign[0]);
+            if (res.imageUrls?.images) setImagesUrl([...res.imageUrls.images, ...imagesUrl]);
+            toast.success('Images uploaded successfully');
         } catch (err) {
             console.log(err);
         }
@@ -60,7 +114,11 @@ const ProductUpdate = () => {
         e.preventDefault();
         try {
             const formData = new FormData();
-            formData.append("image", image);
+            formData.append("frontImage", frontImageUrl);
+            formData.append("backImage", backImageUrl);
+            formData.append("frontDesign", frontDesignUrl);
+            formData.append("backDesign", backDesignUrl);
+            formData.append("images", JSON.stringify(imagesUrl));
             formData.append("name", name);
             formData.append("description", description);
             formData.append("price", price);
@@ -70,14 +128,26 @@ const ProductUpdate = () => {
             formData.append("offers", offers);
             formData.append("returnpolicy", returnpolicy);
 
+            console.log("productData");
+            for (let [key, value] of formData.entries()) {
+                console.log(`${key}:`, value);
+            }
+
             const data = await updateProduct({ productId: productId, formData });
 
+            console.log('data:', data);
             if (data?.error) {
                 console.log(data.error);
-            } else {
+            } else if (data.data.success) {
                 console.log("Product updated successfully", data);
                 router.replace('/ProductList');
             }
+            console.log('front image:', frontImageUrl);
+            console.log('back image:', backImageUrl);
+            console.log('front design:', frontDesignUrl);
+            console.log('back design:', backDesignUrl);
+            console.log('images:', imagesUrl);
+
         } catch (err) {
             console.log(err);
         }
@@ -90,12 +160,48 @@ const ProductUpdate = () => {
             );
             if (!answer) return;
 
-            const { data } = await deleteProduct(productId);
-            console.log('Product deleted successfully', data);
-            router.replace('/ProductList');
+            console.log('Removing product from cart:', productId);
+            const removeFromAllWishlistResponse = await removeFromAllWishlist(productId).unwrap();
+            const removeFromAllCartResponse = await removeAllOfProductFromAllofCart({ productId: productId }).unwrap();
+            if (removeFromAllCartResponse.success && removeFromAllWishlistResponse.success) {
+                const { data } = await deleteProduct(productId);
+                console.log('Product deleted successfully', data);
+                router.replace('/ProductList');
+            }
         } catch (err) {
             console.log(err);
         }
+    };
+
+    const uploadFrontImageHandler = async (e) => {
+        const file = e.target.files[0];
+        setFrontImage(file);
+        setFrontImageUrlToDisplay(URL.createObjectURL(file));
+    };
+
+    const uploadBackImageHandler = async (e) => {
+        const file = e.target.files[0];
+        setBackImage(file);
+        setBackImageUrlToDisplay(URL.createObjectURL(file));
+    };
+
+    const uploadFrontDesignHandler = async (e) => {
+        const file = e.target.files[0];
+        setFrontDesign(file);
+        setFrontDesignUrlToDisplay(URL.createObjectURL(file));
+    };
+
+    const uploadBackDesignHandler = async (e) => {
+        const file = e.target.files[0];
+        setBackDesign(file);
+        setBackDesignUrlToDisplay(URL.createObjectURL(file));
+    };
+
+    const uploadImagesHandler = async (e) => {
+        const files = Array.from(e.target.files);
+        setImages((prevImages) => [...prevImages, ...files]);
+        const urls = files.map((file) => URL.createObjectURL(file));
+        setImagesUrlToDisplay((prevImages) => [...prevImages, ...urls]);
     };
 
     if (!productId) return <div>Loading...</div>;
@@ -104,26 +210,103 @@ const ProductUpdate = () => {
         <div className="create-product-container">
             <h2 className="create-product-heading">Update/Delete Product</h2>
 
-            {image && (
-                <div className="create-product-image">
-                    <img
-                        src={image}
-                        alt="product"
-                    />
-                </div>
-            )}
+            <div className="create-product-images-display">
+                {frontImageUrlToDisplay && (
+                    <div className="create-product-image">
+                        <img
+                            src={frontImageUrlToDisplay}
+                            alt="product"
+                        />
+                        <p>FRONT IMAGE</p>
+                    </div>
+                )}
+                {backImageUrlToDisplay && (
+                    <div className="create-product-image">
+                        <img
+                            src={backImageUrlToDisplay}
+                            alt="product"
+                        />
+                        <p>BACK IMAGE</p>
+                    </div>
+                )}
+                {frontDesignUrlToDisplay && (
+                    <div className="create-product-image">
+                        <img
+                            src={frontDesignUrlToDisplay}
+                            alt="product"
+                        />
+                        <p>FRONT DESIGN</p>
+                    </div>
+                )}
+                {backDesignUrlToDisplay && (
+                    <div className="create-product-image">
+                        <img
+                            src={backDesignUrlToDisplay}
+                            alt="product"
+                        />
+                        <p>BACK DESIGN</p>
+                    </div>
+                )}
+                {imagesUrlToDisplay.map((url, index) => (
+                    <div key={index} className="create-product-image">
+                        <img
+                            src={url}
+                            alt="product"
+                        />
+                        <p>IMAGE {index + 1}</p>
+                    </div>
+                ))}
+            </div>
 
             <div className="create-product-image-upload">
                 <label>
-                    Upload Image
+                    Front Image
                     <input
                         type="file"
                         name="image"
                         accept="image/*"
-                        onChange={uploadFileHandler}
+                        onChange={uploadFrontImageHandler}
+                    />
+                </label>
+                <label>
+                    Back Image
+                    <input
+                        type="file"
+                        name="image"
+                        accept="image/*"
+                        onChange={uploadBackImageHandler}
+                    />
+                </label>
+                <label>
+                    Front Design
+                    <input
+                        type="file"
+                        name="image"
+                        accept="image/*"
+                        onChange={uploadFrontDesignHandler}
+                    />
+                </label>
+                <label>
+                    Back Design
+                    <input
+                        type="file"
+                        name="image"
+                        accept="image/*"
+                        onChange={uploadBackDesignHandler}
+                    />
+                </label>
+                <label>
+                    Images
+                    <input
+                        type="file"
+                        name="image"
+                        accept="image/*"
+                        multiple
+                        onChange={uploadImagesHandler}
                     />
                 </label>
             </div>
+            <button id='create-product-submit' onClick={uploadImages}>Upload Images</button>
 
             <div className="create-product-details">
                 <div className="create-product-details-row1">
